separate mem_utils
