+++ tests/utils.py
@@ -49,7 +49,7 @@ from vllm.utils import (
     cuda_device_count_stateless,
     get_open_port,
 )
-from vllm.utils.mem_utils import GB_bytes
+from vllm.utils.mem_constants import GB_bytes
 
 if current_platform.is_rocm():
     from amdsmi import (
