+++ vllm/utils/mem_utils.py
@@ -10,18 +10,7 @@ from functools import cache
 import psutil
 import torch
 import torch.types
-
-MB_bytes = 1_000_000
-"""The number of bytes in one megabyte (MB)."""
-
-MiB_bytes = 1 << 20
-"""The number of bytes in one mebibyte (MiB)."""
-
-GB_bytes = 1_000_000_000
-"""The number of bytes in one gigabyte (GB)."""
-
-GiB_bytes = 1 << 30
-"""The number of bytes in one gibibyte (GiB)."""
+from mem_constants import GiB_bytes
 
 
 @cache
